refactor: 패키지 구조 변경
/domain -> /model/domain
/dto -> /model/dto
/service -> /model/service
