+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.service.LottoService;
+import lotto.model.service.LottoService;
 import lotto.view.OutputView;
 
 public class Application {
