+++ src/main/java/lotto/model/domain/Lottos.java
@@ -1,8 +1,8 @@
-package lotto.domain;
+package lotto.model.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.dto.LottosDto;
+import lotto.model.dto.LottosDto;
 
 public class Lottos {
