+++ src/main/java/lotto/model/domain/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model.domain;
 
 public enum Rank {
     NOTHING(0),
