+++ src/test/java/lotto/model/domain/WinningNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model.domain;
 
 
 import org.assertj.core.api.Assertions;
